update - refactor getAllItems fetch - add helper functions
